requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
spacy==3.7.2
python-docx==1.0.1
docxtpl==0.16.7
nltk==3.8.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tqdm==4.66.1
openpyxl==3.1.2
jinja2==3.1.2
flask==2.3.3
flask-wtf==1.2.1
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2 